refactor(scenarios): address CodeRabbit review on the judge-transport tests
Extract the stub chat-completions harness to a support module (the test
file exceeded the 300-source-line SRP limit), drop idle keep-alive
sockets via closeAllConnections before resolving the stub's close so
afterEach never waits on them, and rename carriesTools to hasTools per
the boolean-prefix rule.
